+++ src/Structure/Product.java
@@ -12,7 +12,7 @@ public class Product {
 	private int amount_real;
 	private int amount_virtual;
 
-	public Product (String name, float price, String date, String provider, int amount) {
+	public Product (String name, double price, String date, String provider, int amount) {
 		this.name = name;
 		this.price = price;
 		this.expiration = date;
@@ -31,8 +31,8 @@ public class Product {
 	public String getAmountRealAsStr() { return Integer.toString(amount_real); }
 	public String getAmountVirtualAsStr() { return Integer.toString(amount_virtual); }
 
-	public void setPrice(String value) { price = Float.parseFloat(value); }
-	public void setPrice(Float value) { price = value; }
+	public void setPrice(String value) { price = Double.parseDouble(value); }
+	public void setPrice(Double value) { price = value; }
 	public void setExpiration(String value) { expiration = value; }
 	public void setProvider(String value) { provider = value; }
